Changed User and Main.
User: I added a function to get the player's level and did a little bit
of organizing. I am now using toast to dynamically tell the user when
they will level up in any activity.
Main: Changed a bit of what each button does. Level up doesn't use
testactivity anymore (and may never need it again). Also changed main to
dynamically display the current exp and level, but it is ONLY updated
whenever you move to another activity and back. This should be fine
because you shouldn't be gaining exp on the main screen anyway. And even
if you could by walking and such, I believe it should update when the
screen is unlocked again (THIS NEEDS TESTING)

Also, main's screen is now locked to portrait.