Added ListView
I added a simple ListView. Right now, it only displays the title,
description, and deadline for each quest in the list view. I need to add
the other information and make the layout look nicer.

Problem is that my parcel between two activities does not work. So I
cannot add the quests I created from my createToDo activity to the main
activity that displays the quest.